fix: audit polish — DRY CI tag logic, tighten dep bounds, add uv cache
- Extract repeated TAG resolution in update-interpretability.yml into a
  single step with GITHUB_OUTPUT (was duplicated 3 times)
- Add uv dependency caching to python-tests.yml CI workflow
- Tighten numpy upper bound (<3.0 → <2.5) and pyarrow (<24.0 → <24.0.1)
- Remove unnecessary string-quoted type annotations in main.py
- Add explicit dtype=np.float64 to PRODUCTION_BIASES for consistency
- Clarify features.py docstring shape breakdown (25 + 7 + 1)
- Use full-precision bias values in CLAUDE.md

+++ rocket_classifier/model.py
@@ -98,7 +98,9 @@ _GLOBAL_CLASS_PRIOR: np.ndarray = np.array(
 
 # Production threshold biases found via OOB optimisation on Colab.
 # Applied as: preds = argmax(log(proba) + BIASES)
-PRODUCTION_BIASES: np.ndarray = np.array([0.0, -0.25316455696202533, 1.2658227848101262])
+PRODUCTION_BIASES: np.ndarray = np.array(
+    [0.0, -0.25316455696202533, 1.2658227848101262], dtype=np.float64
+)
 
 
 # ---------------------------------------------------------------------------
